+++ src/RequestDataTransformer/RequestDataTransformerInterface.php
@@ -4,12 +4,14 @@ declare(strict_types=1);
 
 namespace DigitalCraftsman\CQRS\RequestDataTransformer;
 
+use DigitalCraftsman\CQRS\Command\Command;
+use DigitalCraftsman\CQRS\Query\Query;
+
 /**
  * The data transformer can have three kinds of tasks and multiple data transformers can be used with one request.
  * - Cast existing data into other formats.
  * - Sanitize existing data.
  * - Add additional data not present in the request.
-
  *  It must not be used to:
  * - Validate the request data in any way. That must be handled in the DTO validator.
  *
@@ -18,6 +20,16 @@ namespace DigitalCraftsman\CQRS\RequestDataTransformer;
  */
 interface RequestDataTransformerInterface
 {
-    /** @param class-string $dtoClass */
-    public function transformRequestData(string $dtoClass, array $requestData): array;
+    /**
+     * @param class-string<Command|Query>               $dtoClass
+     * @param scalar|array<array-key, scalar|null>|null $parameters
+     */
+    public function transformRequestData(
+        string $dtoClass,
+        array $requestData,
+        mixed $parameters,
+    ): array;
+
+    /** @param scalar|array<array-key, scalar|null>|null $parameters */
+    public static function areParametersValid(mixed $parameters): bool;
 }