+++ src/ValueObject/Exception/InvalidParametersInRoutePayload.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCraftsman\CQRS\ValueObject\Exception;
+
+/**
+ * @psalm-immutable
+ *
+ * @codeCoverageIgnore
+ */
+final class InvalidParametersInRoutePayload extends \InvalidArgumentException
+{
+    public function __construct(mixed $class)
+    {
+        parent::__construct(sprintf('Invalid parameters for class %s in route payload', $class));
+    }
+}