+++ tests/Test/Application/UserIdValidator.php
@@ -14,17 +14,27 @@ use Symfony\Component\HttpFoundation\Request;
 final class UserIdValidator implements DTOValidatorInterface
 {
     public function __construct(
-        private SecuritySimulator $securitySimulator,
+        private readonly SecuritySimulator $securitySimulator,
     ) {
     }
 
-    /** @param CreateNewsArticleCommand $dto */
+    /**
+     * @param CreateNewsArticleCommand $dto
+     * @param null                     $parameters
+     */
     public function validateDTO(
         Request $request,
         Command|Query $dto,
+        mixed $parameters,
     ): void {
         if ($this->securitySimulator->getAuthenticatedUserId()->isNotEqualTo($dto->userId)) {
             throw new \DomainException('Supplied user id is invalid');
         }
+    }
+
+    /** @param null $parameters */
+    public static function areParametersValid(mixed $parameters): bool
+    {
+        return $parameters === null;
     }
 }